Fix the sidenav loading (#543)
When a new page was loaded via an HTMX request the interactions with the
sidenav was broken. This was due to the content of "sidenav.js" not
running. The fact that the code was not running was due to the use of
module which are loaded only once.

By exporting the function and calling them explicitly inside the layout
page we can be sure that the code will be always executed.